+++ src/freva/cli/plugin.py
@@ -174,4 +174,5 @@ def main(argv: Optional[list[str]] = None) -> None:
     try:
         cli.run_cmd(args, **cli.kwargs)
     except KeyboardInterrupt:  # pragma: no cover
-        sys.exit(1)
+        print("KeyboardInterrupt, exiting", file=sys.stderr, flush=True)
+        sys.exit(130)